+++ src/Result.cpp
@@ -2,13 +2,13 @@
 
 using namespace std;
 
-Result::Result(int index)
+Result::Result(size_t index)
 {
     lastMatchedIndex = index;
     setOutputs(map<string, forward_list<MatchableInterface *>>());
 }
 
-Result::Result(int index, const map<string, forward_list<MatchableInterface *>> &outputs)
+Result::Result(size_t index, const map<string, forward_list<MatchableInterface *>> &outputs)
 {
     lastMatchedIndex = index;
     setOutputs(outputs);
