+++ src/Result.hpp
@@ -12,8 +12,8 @@ class Result
 {
 public:
     Result();
-    Result(int index);
-    Result(int index, const map<string, forward_list<MatchableInterface *>> &outputs);
+    Result(size_t index);
+    Result(size_t index, const map<string, forward_list<MatchableInterface *>> &outputs);
     ~Result();
 
     bool operator!=(const Result& other) const;
